Correction inventaire
-Désormais, seul le joueur possède un inventaire
-l'inventaire est simplement un tab dynamique d'Objects*
-Suite à une erreur de manipulation git, il se peut qu'il y ai des soucis avec cette branche même si j'ai essayé de tout corriger...

+++ Source/Element/Object/ThrowedObject.cpp
@@ -1,7 +1,7 @@
 #include "ThrowedObject.h"
 #include "iostream"
 
-ThrowedObject::ThrowedObject(std::string texturePath, sf::Vector2f throwerPosition, sf::Vector2f direction, int range, float damages) :
+ThrowedObject::ThrowedObject(std::string texturePath, sf::Vector2f throwerPosition, sf::Vector2f direction, float range, float damages) :
 	Object(texturePath, throwerPosition)
 {
 	_clock = sf::Clock();
